+++ pom.xml
@@ -36,7 +36,7 @@
     <inceptionYear>2015</inceptionYear>
 
     <properties>
-        <main-class>nl.knaw.dans.easy.pid.PidGeneratorService</main-class>
+        <main-class>nl.knaw.dans.easy.pid.service.PidGeneratorService</main-class>
     </properties>
 
     <scm>
@@ -138,6 +138,7 @@
             <version>3.16.1</version>
         </dependency>
     </dependencies>
+
     <repositories>
         <repository>
             <id>DANS</id>
@@ -205,7 +206,7 @@
                         <artifactId>exec-maven-plugin</artifactId>
                         <configuration>
                             <classpathScope>test</classpathScope>
-                            <mainClass>nl.knaw.dans.easy.pid.PidGeneratorService</mainClass>
+                            <mainClass>nl.knaw.dans.easy.pid.service.PidGeneratorService</mainClass>
                         </configuration>
                     </plugin>
                 </plugins>
